{"schema":"decodex/commit/1","summary":"Modularize backend macOS region capture","authority":"manual"}
